+++ src/endpoints/basket/createItem.ts
@@ -5,7 +5,7 @@ import {
   BasketItemDisplayData,
   ItemGroup,
 } from '../../endpoints/basket/getBasket';
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 
 export interface CreateBasketItemParameters {
   basketKey: string;
