+++ src/endpoints/categories/categoryById.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 import {BapiCategory} from '../../types/BapiCategory';
 import {
   CategoryEndpointsParameters,
