+++ src/endpoints/redirects/redirects.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 import {Pagination} from '../products/productsByIds';
 
 export interface GetRedirectsEndpointParameters {
