+++ src/endpoints/search/mappings.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 
 export interface SearchMappingsEndpointParameters {
   term: string;
