+++ src/endpoints/search/resolve.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 
 export interface SearchResolveEndpointParameters {
   term: string;
