+++ src/endpoints/shopconfiguration/shopconfiguration.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 
 export interface ShopConfigurationResponseData {
   shopId: number;
