+++ src/endpoints/typeahead/typeahead.ts
@@ -1,4 +1,4 @@
-import {BapiCall} from '../../interfaces/BapiCall';
+import {BapiCall} from '../../helpers/execute';
 import {
   ProductWith,
   productWithQueryParameterValues,
