+++ src/helpers/__tests__/BapiClient-shopID-query.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../helpers/BapiClient';
+import {StorefrontAPIClient} from '../../BapiClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
